requirements.txt
fastapi
uvicorn
websockets
opencv-python
mediapipe
numpy
google-generativeai
pillow
python-dotenv
pytest
pytest-asyncio
httpx
